+++ Content/src/Database/migrations/2019_09_20_000000_create_posts_table.php
@@ -1,33 +1,34 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-
     public function up()
     {
-        Schema::create('posts', function ($table) {
-            $table->engine = 'InnoDB';
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->nullable()->index();
-            $table->string('title')->nullable();
-            $table->string('slug')->index();
-            $table->text('excerpt')->nullable();
-            $table->longText('content')->nullable();
-            $table->longText('content_html')->nullable();
-            $table->timestamp('published_at')->nullable();
-            $table->boolean('published')->default(false);
-            $table->timestamps();
-            $table->softDeletes();
+        if (! Schema::hasTable('posts')) {
+            Schema::create('posts', function ($table) {
+                $table->engine = 'InnoDB';
+                $table->increments('id');
+                $table->integer('user_id')->unsigned()->nullable()->index();
+                $table->string('title')->nullable();
+                $table->string('slug')->index();
+                $table->text('excerpt')->nullable();
+                $table->longText('content')->nullable();
+                $table->longText('content_html')->nullable();
+                $table->timestamp('published_at')->nullable();
+                $table->boolean('published')->default(false);
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::drop('posts');
+        Schema::dropIfExists('posts');
     }
 };
